update geographiclib usage
