+++ lib/gis/area_poly1.c
@@ -13,6 +13,7 @@
 
 #include <math.h>
 #include <grass/gis.h>
+#include <geodesic.h>
 #include "pi.h"
 
 #define TWOPI M_PI + M_PI
@@ -26,10 +27,13 @@ static struct state {
     double Qp; /** Q at the north pole */
 
     double E; /** Area of the earth */
+
+    struct geod_geodesic g; /** GeographicLib */
 } state;
 
 static struct state *st = &state;
 
+#if 0
 static double Q(double x)
 {
     double sinx, sinx2;
@@ -51,6 +55,7 @@ static double Qbar(double x)
            (st->QbarA +
             cosx2 * (st->QbarB + cosx2 * (st->QbarC + cosx2 * st->QbarD)));
 }
+#endif
 
 /*!
  * \brief Begin area calculations.
@@ -64,8 +69,17 @@ static double Qbar(double x)
  */
 void G_begin_ellipsoid_polygon_area(double a, double e2)
 {
+    double f;
+
+    /* GeographicLib */
+    G_get_ellipsoid_parameters(&a, &e2);
+    f = 1.0 - sqrt(1.0 - e2);
+    geod_init(&st->g, a, f);
+
+#if 0
     double e4, e6;
 
+    /* GRASS native */
     e4 = e2 * e2;
     e6 = e4 * e2;
 
@@ -84,6 +98,7 @@ void G_begin_ellipsoid_polygon_area(double a, double e2)
     st->E = 4 * M_PI * st->Qp * st->AE;
     if (st->E < 0.0)
         st->E = -st->E;
+#endif
 }
 
 /*!
@@ -132,12 +147,29 @@ void G_begin_ellipsoid_polygon_area(double a, double e2)
  */
 double G_ellipsoid_polygon_area(const double *lon, const double *lat, int n)
 {
+    double area;
+
+    /* GeographicLib */
+    struct geod_polygon p;
+    double pP;
+    int i;
+
+    geod_polygon_init(&p, FALSE);
+    /* GeographicLib does not need a closed ring,
+     * see example for geod_polygonarea() in geodesic.h */
+    /* add points in reverse order */
+    i = n;
+    while (--i)
+        geod_polygon_addpoint(&st->g, &p, (double)lat[i], (double)lon[i]);
+    geod_polygon_compute(&st->g, &p, FALSE, TRUE, &area, &pP);
+
+#if 0
     double x1, y1, x2, y2, dx, dy;
     double Qbar1, Qbar2;
-    double area;
-    double thresh =
-        1e-6; /* threshold for dy, should be between 1e-4 and 1e-7 */
+    /* threshold for dy, should be between 1e-4 and 1e-7 */
+    double thresh = 1e-6;
 
+    /* GRASS native */
     x2 = Radians(lon[n - 1]);
     y2 = Radians(lat[n - 1]);
     Qbar2 = Qbar(y2);
@@ -193,6 +225,7 @@ double G_ellipsoid_polygon_area(const double *lon, const double *lat, int n)
         area = st->E;
     if (area > st->E / 2)
         area = st->E - area;
+#endif
 
     return area;
 }
